requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
Pillow==10.4.0
pydantic==2.9.2
python-magic==0.4.27
aiofiles==24.1.0
SQLAlchemy[asyncio]==2.0.36
asyncmy==0.2.10
cryptography==43.0.3
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
